refactor(canvas-editor): split CanvasEditor into folder with hooks
Break the 1344-line CanvasEditor.tsx into a folder:
- PageWrapper.tsx (memoized per-page renderer)
- types.ts (CanvasEditorProps, PageWrapperProps)
- hooks/: useMobileWebViewport, usePageRefs, useLoadedConfigs,
  usePageScrollSync, usePageUndoRedoShortcuts, useToolbarHandlers

Public API (named + default CanvasEditor export) unchanged; tightly
coupled sidebar/subsection-bridge logic kept inline. No behavior change.
